% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ad_util.R
\name{h5ad2data.frame}
\alias{h5ad2data.frame}
\title{Extracts data.frame from h5ad file}
\usage{
h5ad2data.frame(filename, name)
}
\arguments{
\item{filename}{file name or H5IdComponent to read data.frame from}

\item{name}{name of group that contains data.frame to be loaded}
}
\value{
a data.frame
}
\description{
Extracts data.frame from h5ad file
}
\examples{
obs = h5ad2data.frame('adata.h5ad','obs')
}
