% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{h5ad2list}
\alias{h5ad2list}
\title{Loads main parts of h5ad file as ordinary R list}
\usage{
h5ad2list(filename, use.raw = FALSE, load.obsm = FALSE)
}
\arguments{
\item{filename}{path to h5ad file}

\item{load.obsm}{logical, whether to load adata.obsm.}

\item{load.raw}{logical, whether to load adata.raw.X instead of adata.X}
}
\value{
list with following elements:
obs - data.frame
var - data.frame
X - matrix (dense or sparse - as it was in X)
obsm - list of matrices from obsm (presumably reduced dimensions)
}
\description{
Loads main parts of h5ad file as ordinary R list
}
