% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{h5ad2sce}
\alias{h5ad2sce}
\title{Loads h5ad scanpy single cell file as SingleCellExperiment}
\usage{
h5ad2sce(filename, use.raw = FALSE)
}
\arguments{
\item{filename}{path to h5ad file}

\item{use.raw}{logical, whether to attempt to get data from adata.raw. Throws warning (not exception!) if adata.raw is not present and proceeds with adata.X.}
}
\value{
SingleCellExperiment object
}
\description{
functions exports X,obs,var and obsm
it is very experimental and comes with no warranty
}
\examples{
sce = h5ad2sce('adata.h5ad')
}
