% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{h5ad2seurat}
\alias{h5ad2seurat}
\title{Loads h5ad as Seurat object}
\usage{
h5ad2seurat(filename, use.raw = FALSE, load.obsm = TRUE, assay = "RNA")
}
\arguments{
\item{filename}{path to h5ad file}

\item{use.raw}{logical, whether to use adata.raw instead of adata.X}

\item{load.obsm}{logical, whether to load adata.obsm. All matrix-like objects from there will be added as DimReduc to the output object with names coerced to Seurat style (that is not underscores in the middle, single underscore at the end).}

\item{assay}{what assay to put data it (RNA by default)}
}
\value{
Seurat object
}
\description{
For non-spatial data only, use h5ad2seurat_spatial for Visium
}
\examples{
seu = h5ad2seurat('adata.h5ad')
}
