% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{h5ad2seurat_l_spatial}
\alias{h5ad2seurat_l_spatial}
\title{Loads scanpy h5ad file with Visium data into list of Seurat objects}
\usage{
h5ad2seurat_l_spatial(
  filename,
  library_id_field = NULL,
  simplify = TRUE,
  use.raw = FALSE
)
}
\arguments{
\item{filename}{character, path to h5ad file}

\item{library_id_field}{name of adata.obs that specifies visium library. Function trys to guess it if set to NULL (default).}

\item{simplify}{logical, whether return just seurat object (instead of list) if there is only one sample}

\item{use.raw}{logical, whether to attempt to get data from adata.raw. Throws warning (not exception!) if adata.raw is not present and proceeds with adata.X.}
}
\value{
list of Seurat object
}
\description{
simplifies to single Seurat object if there is just one sample (see simplify parameter)
}
