% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{h5ad2seurat_spatial}
\alias{h5ad2seurat_spatial}
\title{Loads scanpy h5ad file with Visium data into Seurat object}
\usage{
h5ad2seurat_spatial(
  filename,
  use.raw = FALSE,
  load.obsm = TRUE,
  simplify = TRUE,
  img.res = "lowres",
  load.X = TRUE
)
}
\arguments{
\item{filename}{character, path to h5ad file}

\item{use.raw}{logical, whether to attempt to get data from adata.raw. Throws warning (not exception!) if adata.raw is not present and proceeds with adata.X.}

\item{load.obsm}{logical, whether to load adata.obsm. All matrix-like objects from there will be added as DimReduc to the output object with names coerced to Seurat style (that is not underscores in the middle, single underscore at the end).}

\item{simplify}{logical, whether to merge loaded samples into single object, return list otherwise.}

\item{img.res}{which of lowres' or 'hires' image to be loaded}

\item{load.X}{logical, whether to load expression. Set all expressions to zeroes if FALSE. It can be much faster not to load expression}
}
\value{
list of Seurat object
}
\description{
Loads scanpy h5ad file with Visium data into Seurat object
}
\examples{
vs = h5ad2seurat_spatial('adata.h5ad')
}
