% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{h5ad2sce}
\alias{h5ad2sce}
\title{Loads h5ad scanpy single cell file as SingleCellExperiment}
\usage{
h5ad2sce(filename)
}
\arguments{
\item{filename}{path to h5ad file}
}
\value{
SingleCellExperiment object
}
\description{
functions exports X,obs,var and obsm
it is very experimental and comes with no warranty
}
\examples{
sce = h5ad2sce('adata.h5ad')
}
