% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{h5ad2seurat_l_spatial}
\alias{h5ad2seurat_l_spatial}
\title{Loads scanpy h5ad file with multiple Visium data into list of Seurat object}
\usage{
h5ad2seurat_l_spatial(filename, library_id_field = NULL)
}
\arguments{
\item{filename}{character, path to h5ad file}

\item{library_id_field}{name of adata.obs that specifies visium library. Function trys to guess it if set to NULL (default).}
}
\value{
list of Seurat object
}
\description{
Loads scanpy h5ad file with multiple Visium data into list of Seurat object
}
