% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ad_util.R
\name{parseH5ADdataframe}
\alias{parseH5ADdataframe}
\title{Extracts dataframe from H5AD list
function is used in conjunction with rhdf5::H5Fopen from myLoadH5AD_Spatials function}
\usage{
parseH5ADdataframe(file, name)
}
\arguments{
\item{collist}{list read from H5AD object}

\item{attr}{}
}
\value{
a data.frame
}
\description{
Extracts dataframe from H5AD list
function is used in conjunction with rhdf5::H5Fopen from myLoadH5AD_Spatials function
}
